[backend] Move and rename cleanMarkings to domain/markingDefinition
